requirements.txt
python-dotenv
google-genai
mem0ai
streamlit